requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
torch==2.5.1
torchvision==0.20.1
Pillow==10.2.0
werkzeug==3.0.1
